streamlit_practice/pyproject.toml
[project]
name = "streamlit-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.1.0",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "langchain-pinecone>=0.2.13",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
